Update: community && communityComment CRUD

+++ src/components/mypage/mypageFNs.ts
@@ -8,14 +8,21 @@ export const getMyInfo = async () => {
 };
 
 export const deleteUser = async () => {
-  try {
-    await userAPI.deleteUser();
-    swalSuccess("회원탈퇴가 완료되었습니다.").then(() => {
-      goHome();
-    });
-  } catch (error) {
-    swalError("알 수 없는 오류입니다.");
-  }
+  swalQuestion(
+    "정말 탈퇴할까요?",
+    "삭제되는 유저 정보, 게시글, 댓글 등은 복구되지 않습니다.",
+  ).then(async res => {
+    if (res.value) {
+      try {
+        await userAPI.deleteUser();
+        swalSuccess("회원탈퇴가 완료되었습니다.").then(() => {
+          goHome();
+        });
+      } catch (error) {
+        swalError("알 수 없는 오류입니다.");
+      }
+    }
+  });
 };
 
 export const deleteCheck = () => {
